Delete generateCategoricalChart.tsx
