+++ scripts/snapshots/es6Files.txt
@@ -35,7 +35,6 @@
   "es6/chart/ScatterChart.js",
   "es6/chart/SunburstChart.js",
   "es6/chart/Treemap.js",
-  "es6/chart/generateCategoricalChart.js",
   "es6/chart/types.js",
   "es6/component",
   "es6/component/ActivePoints.js",
