+++ scripts/snapshots/libFiles.txt
@@ -35,7 +35,6 @@
   "lib/chart/ScatterChart.js",
   "lib/chart/SunburstChart.js",
   "lib/chart/Treemap.js",
-  "lib/chart/generateCategoricalChart.js",
   "lib/chart/types.js",
   "lib/component",
   "lib/component/ActivePoints.js",
