+++ scripts/snapshots/typesFiles.txt
@@ -35,7 +35,6 @@
   "types/chart/ScatterChart.d.ts",
   "types/chart/SunburstChart.d.ts",
   "types/chart/Treemap.d.ts",
-  "types/chart/generateCategoricalChart.d.ts",
   "types/chart/types.d.ts",
   "types/component",
   "types/component/ActivePoints.d.ts",
