+++ src/chart/BarChart.tsx
@@ -1,8 +1,7 @@
 import * as React from 'react';
 import { forwardRef } from 'react';
-import { CategoricalChartProps } from './generateCategoricalChart';
 import { arrayTooltipSearcher } from '../state/optionsSlice';
-import { TooltipEventType } from '../util/types';
+import { CategoricalChartProps, TooltipEventType } from '../util/types';
 import { CartesianChart } from './CartesianChart';
 
 const allowedTooltipTypes: ReadonlyArray<TooltipEventType> = ['axis', 'item'];
