+++ src/chart/LineChart.tsx
@@ -2,8 +2,7 @@ import * as React from 'react';
 import { forwardRef } from 'react';
 import { arrayTooltipSearcher } from '../state/optionsSlice';
 import { CartesianChart } from './CartesianChart';
-import { CategoricalChartProps } from './generateCategoricalChart';
-import { TooltipEventType } from '../util/types';
+import { CategoricalChartProps, TooltipEventType } from '../util/types';
 
 const allowedTooltipTypes: ReadonlyArray<TooltipEventType> = ['axis'];
