+++ src/state/selectors/selectActivePropsFromChartPointer.ts
@@ -1,12 +1,12 @@
 import { createSelector } from 'reselect';
 import { RechartsRootState } from '../store';
-import { ChartPointer } from '../../chart/generateCategoricalChart';
 import { ActiveTooltipProps } from '../tooltipSlice';
 import { selectChartLayout } from '../../context/chartLayoutContext';
 import { selectTooltipAxisRangeWithReverse, selectTooltipAxisTicks, selectTooltipAxisType } from './tooltipSelectors';
 import { selectChartOffset } from './selectChartOffset';
 import { combineActiveProps, selectOrderedTooltipTicks } from './selectors';
 import { selectPolarViewBox } from './polarAxisSelectors';
+import { ChartPointer } from '../../util/types';
 
 const pickChartPointer = (_state: RechartsRootState, chartPointer: ChartPointer) => chartPointer;
